Move addTranscriptionLikeElement from editor to markup
